Move processors to common directory
