+++ src/lib/controller/index.ts
@@ -10,8 +10,8 @@ import { MutateResponse, ValidateResponse } from "../k8s";
 import Log from "../telemetry/logger";
 import { metricsCollector, MetricsCollector } from "../telemetry/metrics";
 import { ModuleConfig, isWatchMode } from "../module";
-import { mutateProcessor } from "../mutate-processor";
-import { validateProcessor } from "../validate-processor";
+import { mutateProcessor } from "../processors/mutate-processor";
+import { validateProcessor } from "../processors/validate-processor";
 import { StoreController } from "./store";
 import { AdmissionRequest } from "../types";
 import { karForMutate, karForValidate, KubeAdmissionReview } from "./index.util";
