+++ src/lib/processors/watch-processor.test.ts
@@ -4,22 +4,22 @@ import { afterAll, beforeEach, describe, expect, it, jest } from "@jest/globals"
 import { GenericClass, K8s, KubernetesObject, kind } from "kubernetes-fluent-client";
 import { K8sInit, WatchPhase } from "kubernetes-fluent-client/dist/fluent/types";
 import { WatchCfg, WatchEvent, Watcher } from "kubernetes-fluent-client/dist/fluent/watch";
-import { Capability } from "./capability";
+import { Capability } from "../capability";
 import { setupWatch, logEvent, queueKey, getOrCreateQueue } from "./watch-processor";
-import Log from "./telemetry/logger";
-import { metricsCollector } from "./telemetry/metrics";
+import Log from "../telemetry/logger";
+import { metricsCollector } from "../telemetry/metrics";
 
 type onCallback = (eventName: string | symbol, listener: (msg: string) => void) => void;
 
 // Mock the dependencies
 jest.mock("kubernetes-fluent-client");
 
-jest.mock("./telemetry/logger", () => ({
+jest.mock("../telemetry/logger", () => ({
   debug: jest.fn(),
   error: jest.fn(),
 }));
 
-jest.mock("./telemetry/metrics", () => ({
+jest.mock("../telemetry/metrics", () => ({
   metricsCollector: {
     initCacheMissWindow: jest.fn(),
     incCacheMiss: jest.fn(),
