+++ src/lib/processors/watch-processor.ts
@@ -2,14 +2,14 @@
 // SPDX-FileCopyrightText: 2023-Present The Pepr Authors
 import { K8s, KubernetesObject, WatchCfg, WatchEvent } from "kubernetes-fluent-client";
 import { WatchPhase } from "kubernetes-fluent-client/dist/fluent/types";
-import { Capability } from "./capability";
-import { filterNoMatchReason } from "./helpers";
-import { removeFinalizer } from "./finalizer";
-import Log from "./telemetry/logger";
-import { Queue } from "./queue";
-import { Binding } from "./types";
-import { Event } from "./enums";
-import { metricsCollector } from "./telemetry/metrics";
+import { Capability } from "../capability";
+import { filterNoMatchReason } from "../helpers";
+import { removeFinalizer } from "../finalizer";
+import Log from "../telemetry/logger";
+import { Queue } from "../queue";
+import { Binding } from "../types";
+import { Event } from "../enums";
+import { metricsCollector } from "../telemetry/metrics";
 
 // stores Queue instances
 const queues: Record<string, Queue<KubernetesObject>> = {};
